batch_utils
