+++ deepchem/utils/__init__.py
@@ -6,6 +6,8 @@ import logging
 
 logger = logging.getLogger(__name__)
 
+from deepchem.utils.batch_utils import coulomb_matrix_features
+
 from deepchem.utils.conformers import ConformerGenerator
 from deepchem.utils.evaluate import relative_difference
 from deepchem.utils.evaluate import Evaluator
